refactor(auth-*-chat): extract keypair/JWT minting into per-example jwt.ts
The vitest.config.browser.ts files were doing real work at module-load
time — generating an ES256 keypair, deriving a JWK, and minting JWTs —
which made config failures harder to localise. Move that into a small
tests/browser/jwt.ts in each example exporting createTestKeySet() with
{ publicJwk, mintJwt }. Keep the file per-example rather than shared so
each example stays self-contained.

Also add a comment in each config explaining the config → process.env →
global-setup boundary (vitest's only channel from config to globalSetup
is env vars).
