+++ examples/auth-simple-chat/vitest.config.browser.ts
@@ -4,10 +4,8 @@ import wasm from "vite-plugin-wasm";
 import topLevelAwait from "vite-plugin-top-level-await";
 import react from "@vitejs/plugin-react";
 import { playwright } from "@vitest/browser-playwright";
-import { exportJWK, generateKeyPair, SignJWT } from "jose";
+import { createTestKeySet } from "./tests/browser/jwt.js";
 import { TEST_APP_ID } from "./tests/browser/test-constants.js";
-
-const KID = "auth-simple-chat-test-key";
 
 function findFreePort(): Promise<number> {
   return new Promise((resolve, reject) => {
@@ -21,24 +19,16 @@ function findFreePort(): Promise<number> {
 }
 
 export default defineConfig(async () => {
-  const { publicKey, privateKey } = await generateKeyPair("ES256", { extractable: true });
-  const publicJwk = { ...(await exportJWK(publicKey)), kid: KID, use: "sig", alg: "ES256" };
-
-  async function mintJwt(role: string, sub: string): Promise<string> {
-    return new SignJWT({ claims: { role } })
-      .setProtectedHeader({ alg: "ES256", kid: KID })
-      .setSubject(sub)
-      .setIssuedAt()
-      .setExpirationTime("1h")
-      .sign(privateKey);
-  }
-
+  const { publicJwk, mintJwt } = await createTestKeySet();
   const adminJwt = await mintJwt("admin", "admin-test-user");
   const memberJwt = await mintJwt("member", "member-test-user");
 
   const jwksPort = await findFreePort();
   const jazzPort = await findFreePort();
 
+  // Vitest doesn't pass state from config to global-setup directly; the only
+  // channel is process.env. global-setup.ts reads these via requireEnv() to
+  // boot the JWKS server and the Jazz TestingServer.
   process.env.JAZZ_TEST_JWKS_PUBLIC_KEY = JSON.stringify(publicJwk);
   process.env.JAZZ_TEST_JWKS_PORT = String(jwksPort);
   process.env.JAZZ_TEST_JAZZ_PORT = String(jazzPort);
